+++ Node/Messages/HeadersMessage.cs
@@ -36,7 +36,7 @@ namespace BToken
     {
       int startIndex = 0;
 
-      int headersCount = (int)VarInt.GetUInt64(Payload, ref startIndex);
+      int headersCount = VarInt.GetInt32(Payload, ref startIndex);
       for (int i = 0; i < headersCount; i++)
       {
         Headers.Add(NetworkHeader.ParseHeader(Payload, out int txCount, ref startIndex));
